Point imports at features/, and split the duckerweb type file
types/type.ts mixed two concerns. Renderer types (GHNode, Port, Bounds, all
GH*NodeProps, the handle and edge types) move to features/gh-flow/types.ts;
duckerweb keeps its own state, diff, and view-control types.

ParsedComponent stays with duckerweb rather than moving to gh-flow as the
plan suggested - its only consumers are component-list and component-card.

Share now declares its own ShareViewMode ("list" | "flow") instead of
borrowing duckerweb's four-value superset, which had made share look like it
supported a diff view it has no code for.
