+++ src/features/gh-flow/components/gh-panel-node.tsx
@@ -1,6 +1,6 @@
-import type { GHNodeProps } from "../types/type";
-import { GHHandle } from "./Handle";
-import { HandlePosition } from "./HandlePosition";
+import type { GHNodeProps } from "../types";
+import { GHHandle } from "./handle";
+import { HandlePosition } from "./handle-position";
 
 export function GHPanelNode({ data }: GHNodeProps) {
 	const inputs = data.inputs ?? [];
